+++ kifi-backend/modules/common/app/com/keepit/common/cache/FortyTwoCache.scala
@@ -12,7 +12,8 @@ import net.sf.ehcache.config.CacheConfiguration
 
 import com.google.inject.{Inject, Singleton}
 import com.keepit.common.healthcheck.{AirbrakeNotifier, AirbrakeError}
-import com.keepit.common.logging.Logging
+import com.keepit.common.logging.Access.CACHE
+import com.keepit.common.logging._
 import com.keepit.common.time._
 import com.keepit.serializer.{Serializer, BinaryFormat}
 import com.keepit.common.logging.{AccessLogTimer, AccessLog}
@@ -24,242 +25,14 @@ import play.api.libs.concurrent.Execution.Implicits._
 import play.api.libs.json._
 import play.modules.statsd.api.Statsd
 
-//hate having an object here, should refactor to @Singleton
-object GlobalCacheStatistics {
-  private[cache] val hitsMap = ConcurrentMap[String, AtomicInteger]()
-  private[cache] val missesMap = ConcurrentMap[String, AtomicInteger]()
-  private[cache] val setsMap = ConcurrentMap[String, AtomicInteger]()
-
-  def getStatistics: Seq[(String, Int, Int, Int)] = {
-    val keys = (hitsMap.keySet ++ missesMap.keySet ++ setsMap.keySet).toSeq.sorted
-    keys map { key =>
-      (key, getCount(key, hitsMap), getCount(key, missesMap), getCount(key, setsMap))
-    }
-  }
-
-  private[cache] def getCount(key: String, m: ConcurrentMap[String, AtomicInteger]): Int = {
-    m.get(key) match {
-      case Some(counter) => counter.get()
-      case _ => 0
-    }
-  }
-}
-
-trait CacheStatistics extends Logging {
-  val global = GlobalCacheStatistics
-  private def incrCount(key: String, m: ConcurrentMap[String, AtomicInteger]) {
-    m.getOrElseUpdate(key, new AtomicInteger(0)).incrementAndGet()
-  }
-  private val accessLog = Logger("com.keepit.access")
-
-  def recordHit(cachePlugin: String, logAccess: Boolean, namespace: String, fullKey: String, millis: Long) {
-    val name = s"$cachePlugin.$namespace"
-    incrCount(name, global.hitsMap)
-    Statsd.increment(s"$name.hits")
-    Statsd.timing(s"$name.hits", millis)
-    if (logAccess) accessLog.info(s"""[CACHE] [$cachePlugin] HIT  $fullKey took [${millis}ms]""")
-  }
-
-  def recordMiss(cachePlugin: String, logAccess: Boolean, namespace: String, fullKey: String, millis: Long) {
-    val name = s"$cachePlugin.$namespace"
-    incrCount(s"$name", global.missesMap)
-    Statsd.increment(s"$name.misses")
-    log.warn(s"Cache miss on key $fullKey in $cachePlugin")
-    if (logAccess) accessLog.info(s"""[CACHE] [$cachePlugin] MISS $fullKey took [${millis}ms]""")
-  }
-
-  def recordSet(cachePlugin: String, logAccess: Boolean, namespace: String, fullKey: String, millis: Long) {
-    val name = s"$cachePlugin.$namespace"
-    incrCount(s"$name", global.setsMap)
-    Statsd.increment(s"$name.sets")
-    Statsd.timing(s"$name.sets", millis)
-    if (logAccess) accessLog.info(s"""[CACHE] [$cachePlugin] SET  $fullKey took [${millis}ms]""")
-  }
-}
-
-trait FortyTwoCachePlugin extends Plugin {
-  private[cache] def onError(error: AirbrakeError) {}
-
-  private[cache] val logAccess: Boolean = false
-
-  def get(key: String): Option[Any]
-  def remove(key: String): Unit
-  def set(key: String, value: Any, expiration: Int = 0): Unit
-
-  override def enabled = true
-
-  override def toString = "Cache"
-}
-
-trait InMemoryCachePlugin extends FortyTwoCachePlugin
-
-@Singleton
-class MemcachedCache @Inject() (
-  val cache: MemcachedPlugin,
-  val airbrake: AirbrakeNotifier) extends FortyTwoCachePlugin {
-
-  override private[cache] val logAccess = true
-
-  def get(key: String): Option[Any] = cache.api.get(key)
-
-  override def onError(error: AirbrakeError) {
-    airbrake.notify(error)
-  }
-
-  def remove(key: String) = cache.api.remove(key)
-
-  def set(key: String, value: Any, expiration: Int = 0): Unit = future {
-    cache.api.set(key, value, expiration)
-  }
-
-  override def onStop() = cache.onStop()
-
-  override def toString = "Memcached"
-}
-
-class EhCacheConfiguration extends CacheConfiguration
-
-@Singleton
-class EhCacheCache @Inject() (
-  config: EhCacheConfiguration,
-  val airbrake: AirbrakeNotifier)
-    extends InMemoryCachePlugin {
-
-  lazy val (manager, cache) = {
-    val manager = CacheManager.create()
-    val cache = new Cache(config)
-    manager.addCache(cache)
-    (manager, cache)
-  }
-  override def onStart() { cache }
-  override def onStop() { manager. shutdown() }
-  override def onError(error: AirbrakeError) { airbrake.notify(error) }
-
-  def get(key: String): Option[Any] = Option(cache.get(key)).map(_.getObjectValue)
-  def remove(key: String) { cache.remove(key) }
-
-  def set(key: String, value: Any, expiration: Int = 0): Unit = future {
-    val element = new Element(key, value)
-    if (expiration == 0) element.setEternal(true)
-    element.setTimeToLive(expiration)
-    cache.put(element)
-  }
-
-  override def toString = "EhCache"
-
-}
-
-trait Key[T] {
-  val namespace: String
-  val version: Int = 1
-  def toKey(): String
-  override final def toString: String = namespace + "%" + version + "#" + toKey()
-}
-
 case class CacheSizeLimitExceededException(msg:String) extends Exception(msg)
-
-trait ObjectCache[K <: Key[T], T] {
-  val outerCache: Option[ObjectCache[K, T]] = None
-  val ttl: Duration
-  outerCache map {outer => require(ttl <= outer.ttl)}
-
-  protected[cache] def getFromInnerCache(key: K): Option[Option[T]]
-  protected[cache] def setInnerCache(key: K, value: Option[T]): Unit
-
-  def remove(key: K): Unit
-
-  def set(key: K, value: T): Unit = {
-    outerCache map {outer => outer.set(key, value)}
-    setInnerCache(key, Some(value))
-  }
-
-  def set(key: K, valueOpt: Option[T]) : Unit = {
-    outerCache map {outer => outer.set(key, valueOpt)}
-    setInnerCache(key, valueOpt)
-  }
-
-  def get(key: K): Option[T] = {
-    getFromInnerCache(key) match {
-      case Some(valueOpt) => valueOpt
-      case None => outerCache match {
-        case Some(cache) => cache.get(key)
-        case None => None
-      }
-    }
-  }
-
-  def getOrElse(key: K)(orElse: => T): T = {
-    def fallback : T = {
-      val value = outerCache match {
-        case Some(cache) => cache.getOrElse(key)(orElse)
-        case None => orElse
-      }
-      setInnerCache(key, Some(value))
-      value
-    }
-
-    getFromInnerCache(key) match {
-      case Some(valueOpt) => valueOpt match {
-        case Some(value) => value
-        case None => fallback
-      }
-      case None => fallback
-
-    }
-  }
-
-  def getOrElseOpt(key: K)(orElse: => Option[T]): Option[T] = {
-    getFromInnerCache(key) match {
-      case Some(valueOpt) => valueOpt
-      case None =>
-        val valueOption : Option[T] = outerCache match {
-          case Some(cache) => cache.getOrElseOpt(key)(orElse)
-          case None => orElse
-        }
-        setInnerCache(key, valueOption)
-        valueOption
-    }
-  }
-
-  def getOrElseFuture(key: K)(orElse: => Future[T]): Future[T] = {
-    def fallback: Future[T] = {
-      val valueFuture = outerCache match {
-        case Some(cache) => cache.getOrElseFuture(key)(orElse)
-        case None => orElse
-      }
-      valueFuture.onSuccess {case value => setInnerCache(key, Some(value))}
-      valueFuture
-    }
-
-    getFromInnerCache(key) match {
-      case Some(valueOpt) => valueOpt match {
-        case Some(value) => Promise.successful(value).future
-        case None => fallback
-      }
-      case None => fallback
-    }
-  }
-
-  def getOrElseFutureOpt(key: K)(orElse: => Future[Option[T]]): Future[Option[T]] = {
-    getFromInnerCache(key) match {
-      case Some(valueOpt) => Promise.successful(valueOpt).future
-      case None =>
-        val valueFutureOption = outerCache match {
-          case Some(cache) => cache.getOrElseFutureOpt(key)(orElse)
-          case None => orElse
-        }
-        valueFutureOption.onSuccess {case valueOption => setInnerCache(key, valueOption)}
-        valueFutureOption
-    }
-  }
-}
 
 trait FortyTwoCache[K <: Key[T], T] extends ObjectCache[K, T] with CacheStatistics {
   val repo: FortyTwoCachePlugin
   val serializer: Serializer[T]
 
   protected[cache] def getFromInnerCache(key: K): Option[Option[T]] = {
-    val getStart = currentDateTime.getMillis()
+    val timer = accessLog.timer(CACHE)
     val valueOpt = try repo.get(key.toString) catch {
       case e: Throwable =>
         repo.onError(AirbrakeError(e, Some(s"Failed fetching key $key from $repo")))
@@ -267,7 +40,6 @@ trait FortyTwoCache[K <: Key[T], T] extends ObjectCache[K, T] with CacheStatisti
     }
     try {
       val objOpt = valueOpt.map(serializer.reads)
-      val time = currentDateTime.getMillis() - getStart
       objOpt match {
         case Some(_) => {
           recordHit(repo.toString, repo.logAccess, key.namespace, key.toString, time)
