+++ kifi-backend/modules/common/app/com/keepit/common/cache/Key.scala
@@ -1,0 +1,9 @@
+package com.keepit.common.cache
+
+trait Key[T] {
+  val namespace: String
+  val version: Int = 1
+  def toKey(): String
+  override final def toString: String = namespace + "%" + version + "#" + toKey()
+}
+
