+++ kifi-backend/modules/common/app/com/keepit/common/logging/AccessLogEvent.scala
@@ -40,7 +40,7 @@ case class AccessLogTimer(eventType: AccessLogEventType, clock: Clock) {
   //using null for internal api to make the usage of the call much more friendly without having Some(foo) instead of just foo's
   def done(remoteTime: Long = NoLongValue,
           statusCode: Int = NoIntValue,
-          success: Option[Boolean] = None,//can't get away without option here
+          result: String = null,
           remoteHost: String = null,
           targetHost: String = null,
           remoteService: String = null,
@@ -48,6 +48,8 @@ case class AccessLogTimer(eventType: AccessLogEventType, clock: Clock) {
           query: String = null,
           trackingId: String = null,
           method: String = null,
+          key: String = null,
+          space: String = null,
           url: String = null) = {
     val now = clock.now()
     AccessLogEvent(
@@ -56,7 +58,7 @@ case class AccessLogTimer(eventType: AccessLogEventType, clock: Clock) {
       eventType = eventType,
       remoteTime = longOption(remoteTime),
       statusCode = intOption(statusCode),
-      success = success,
+      result = Option(result),
       remoteHost = Option(remoteHost),
       targetHost = Option(targetHost),
       remoteService = Option(remoteService),
@@ -64,6 +66,8 @@ case class AccessLogTimer(eventType: AccessLogEventType, clock: Clock) {
       query = Option(query),
       trackingId = Option(trackingId),
       method = Option(method),
+      key = Option(key),
+      space = Option(space),
       url = Option(url))
   }
 }
@@ -74,7 +78,7 @@ case class AccessLogEvent(
   eventType: AccessLogEventType,
   remoteTime: Option[Long],
   statusCode: Option[Int],
-  success: Option[Boolean],
+  result: Option[String],
   remoteHost: Option[String],
   targetHost: Option[String],
   remoteService: Option[String],
@@ -82,6 +86,8 @@ case class AccessLogEvent(
   query: Option[String],
   trackingId: Option[String],
   method: Option[String],
+  key: Option[String],
+  space: Option[String],
   url: Option[String])
 
 @Singleton
@@ -105,10 +111,12 @@ class AccessLog @Inject() (clock: Clock) {
       Some(s"type:${e.eventType.name}") ::
       Some(s"duration:${e.duration}") ::
       e.method.map("method:" + _) ::
+      e.key.map("key:" + _) ::
+      e.space.map("space:" + _) ::
       e.remoteTime.map("remoteTime:" + _) ::
       e.remoteTime.map(t => "waitTime:" + (e.duration - t)) ::
       e.statusCode.map("statusCode:" + _) ::
-      e.success.map("success:" + _) ::
+      e.result.map("result:" + _) ::
       e.remoteHost.map("remoteHost:" + _) ::
       e.targetHost.map("targetHost:" + _) ::
       e.remoteService.map("remoteService:" + _) ::
